# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VoltHub CSMS API",
    "version": "1.0.0",
    "description": "Two-engine EV charging management. Oracle error bands (-205xx..-208xx) map to HTTP in `error.ora`."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/docs": {
      "get": {
        "security": [],
        "summary": "This OpenAPI document",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/me/vehicles/{id}": {
      "patch": {
        "summary": "Rename / set default EV",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/me/notifications": {
      "get": {
        "summary": "In-app events (reservation/session/invoice)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/me/notifications/{id}/read": {
      "post": {
        "summary": "Mark notification read",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/stations/{id}/sessions/active": {
      "get": {
        "summary": "Active sessions on this station (auth: driver sees own scope, staff scoped)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/stations/{id}/reviews": {
      "get": {
        "security": [],
        "summary": "Station reviews",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/tariffs/active": {
      "get": {
        "security": [],
        "summary": "Current tariff versions + ToU bands (public preview)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/sessions/{id}/state": {
      "patch": {
        "summary": "Operator transition: SUSPENDED↔CHARGING (matrix-enforced)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/admin/stations": {
      "get": {
        "summary": "All stations incl. inactive (staff)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      },
      "post": {
        "summary": "Create station + charge points + connectors (admin)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/admin/stations/{id}": {
      "patch": {
        "summary": "Activate/deactivate (ACTIVE|INACTIVE), reassign operator (admin)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/admin/charge-points": {
      "post": {
        "summary": "Provision charge point → OCPP identity (admin)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
...
    "/ai/forecast": {
      "get": {
        "summary": "AI advisory: per-station hourly demand forecast (operator in scope / admin; 503 when the sidecar is down)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/ai/anomalies": {
      "get": {
        "summary": "AI advisory: robust z-score anomalies over recent station ticks (operator in scope / admin)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    },
    "/ai/optimize": {
      "post": {
        "summary": "AI advisory: deterministic LP smart-charging schedule under hard station/connector constraints (operator in scope / admin)",
        "responses": {
          "200": {
            "description": "ok"
          },
          "4xx": {
            "description": "client error"
          }
        }
      }
    }
  }
}